tipo cambiado (cantidad)

+++ ProgramaApuestasJDBC/src/interfaces/Ingreso.java
@@ -6,7 +6,7 @@ public interface Ingreso {
     public int getId() ;
     public void setId(int id) ;
     public double getCantidad();
-    public void setCantidad(int cantidad);
+    public void setCantidad(double cantidad);
     public String getDescripcion() ;
     public void setDescripcion(String descripcion) ;
     public UsuarioImpl getUsuario() ;
